packages/agent-runtime/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "pnlclaw-agent"
version = "0.1.0"
description = "Strategy drafting and explanation workflows"
readme = {text = "Strategy drafting and explanation workflows", content-type = "text/plain"}
license = "AGPL-3.0-only"
requires-python = ">=3.11"
dependencies = [
    "pnlclaw-types",
    "pnlclaw-core",
    "pnlclaw-security",
    "pnlclaw-market",
    "pnlclaw-strategy",
    "pnlclaw-backtest",
    "pnlclaw-paper",
    "pnlclaw-risk",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
mcp = [
    "mcp>=1.0",
]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "mcp>=1.0",
]

[tool.hatch.build.targets.wheel]
packages = ["pnlclaw_agent"]
